+++ packages/string_literal_finder/bin/string_literal_finder.dart
@@ -20,9 +20,11 @@ const ARG_SILENT = 'silent';
 const ARG_EXCLUDE_PATH = 'exclude-path';
 const ARG_EXCLUDE_SUFFIX = 'exclude-suffix';
 const ARG_METRICS_FILE = 'metrics-output-file';
+const ARG_ANNOTATION_FILE = 'annotations-output-file';
+const ARG_ANNOTATION_ROOT = 'annotations-path-root';
 
 Future<void> main(List<String> arguments) async {
-//  PrintAppender.setupLogging(level: Level.FINER);
+  PrintAppender.setupLogging(level: Level.SEVERE);
   final parser = ArgParser()
     ..addOption(
       ARG_PATH,
@@ -31,6 +33,11 @@ Future<void> main(List<String> arguments) async {
     )
     ..addOption(ARG_METRICS_FILE,
         abbr: 'm', help: 'File to write json metrics to')
+    ..addOption(ARG_ANNOTATION_FILE,
+        help: 'File for annotations as taken by '
+            'https://github.com/Attest/annotations-action/')
+    ..addOption(ARG_ANNOTATION_ROOT,
+        help: 'Maks paths relative to the given root directory.')
     ..addMultiOption(ARG_EXCLUDE_PATH,
         help: 'Exclude paths (relative to path, "startsWith").')
     ..addMultiOption(ARG_EXCLUDE_SUFFIX,
@@ -62,6 +69,11 @@ Future<void> main(List<String> arguments) async {
           .toList(),
     );
     final foundStringLiterals = await stringLiteralFinder.start();
+
+    await (results[ARG_ANNOTATION_FILE] as String)?.let((file) =>
+        _generateAnnotationsFile(file, foundStringLiterals,
+            pathRelativeFrom: results[ARG_ANNOTATION_ROOT] as String));
+
     final fileCount = foundStringLiterals.map((e) => e.filePath).toSet();
     final nonLiteralFiles =
         stringLiteralFinder.filesAnalyzed.difference(fileCount);
@@ -91,6 +103,29 @@ Future<void> main(List<String> arguments) async {
     _logger.severe('Error during analysis.', e, stackTrace);
     exitCode = 70;
   }
+}
+
+Future<void> _generateAnnotationsFile(
+  String file,
+  List<FoundStringLiteral> foundStringLiterals, {
+  String pathRelativeFrom,
+}) async {
+  print('ok?');
+  final pathValue = pathRelativeFrom
+          ?.let((from) => (String p) => path.relative(p, from: from)) ??
+      (String p) => p;
+  final annotations = foundStringLiterals
+      .map(
+        (e) => {
+          'message': 'String literal',
+          'level': 'notice',
+          'path': pathValue(e.filePath),
+          'column': {'start': e.loc.columnNumber, 'end': e.locEnd.columnNumber},
+          'line': {'start': e.loc.lineNumber, 'end': e.locEnd.lineNumber},
+        },
+      )
+      .toList();
+  await File(file).writeAsString(json.encode(annotations));
 }
 
 extension<T extends Object> on T {