+++ packages/string_literal_finder/test/string_literal_finder_test.dart
@@ -1,4 +1,3 @@
-import 'package:string_literal_finder/string_literal_finder.dart';
 import 'package:test/test.dart';
 
 void main() {